nevera/src/train_requirements.txt
numpy==1.17.3
keras==2.6.0
Keras-Preprocessing==1.1.2
pandas==1.1.5
Pillow==8.4.0
scikit-image==0.18.3
tensorflow-cpu
urllib3==1.24.3
imutils==0.5.3
gluoncv==0.7.0

